+++ packages/core-mobile/app/new/features/meldOnramp/components/PaymentMethodIcon.tsx
