+++ packages/core-mobile/app/new/features/meldOnramp/components/ServiceProviderIcon.tsx
