+++ packages/core-mobile/app/new/features/meldOnramp/hooks/useSearchServiceProviders.ts
@@ -1,5 +1,5 @@
 import { useQuery, UseQueryResult } from '@tanstack/react-query'
-import MeldService from 'features/buyOnramp/services/MeldService'
+import MeldService from 'features/meldOnramp/services/MeldService'
 import { ReactQueryKeys } from 'consts/reactQueryKeys'
 import { MeldDefaultParams, ServiceProvider } from '../types'
 import { useLocale } from './useLocale'
