+++ packages/core-mobile/app/new/features/portfolio/assets/screens/TokenDetailScreen.tsx
@@ -63,6 +63,10 @@ import {
 import { selectIsDeveloperMode } from 'store/settings/advanced'
 import { selectSelectedCurrency } from 'store/settings/currency'
 import { selectIsPrivacyModeEnabled } from 'store/settings/securityPrivacy'
+import { useFormatCurrency } from 'common/hooks/useFormatCurrency'
+import { useNetworks } from 'hooks/networks/useNetworks'
+import { ChainId } from '@avalabs/core-chains-sdk'
+import { useBuy } from 'features/meldOnramp/hooks/useBuy'
 
 export const TokenDetailScreen = (): React.JSX.Element => {
   const {
