+++ packages/core-mobile/app/new/features/portfolio/screens/PortfolioScreen.tsx
@@ -22,7 +22,6 @@ import { useSearchableTokenList } from 'common/hooks/useSearchableTokenList'
 import { useIsAndroidWithBottomBar } from 'common/hooks/useIsAndroidWithBottomBar'
 import { UNKNOWN_AMOUNT } from 'consts/amount'
 import { useFocusEffect, useRouter } from 'expo-router'
-import { useBuy } from 'features/buyOnramp/hooks/useBuy'
 import {
   ActionButton,
   ActionButtons
@@ -71,6 +70,7 @@ import { selectSelectedCurrency } from 'store/settings/currency'
 import { selectIsPrivacyModeEnabled } from 'store/settings/securityPrivacy'
 import { RootState } from 'store/types'
 import { useFocusedSelector } from 'utils/performance/useFocusedSelector'
+import { useBuy } from 'features/meldOnramp/hooks/useBuy'
 
 const SEGMENT_ITEMS = ['Assets', 'Collectibles', 'DeFi']
