+++ packages/core-mobile/app/new/features/stake/hooks/useAddStake.ts
@@ -1,7 +1,7 @@
 import { showAlert } from '@avalabs/k2-alpine'
 import { useRouter } from 'expo-router'
-import { useBuy } from 'features/buyOnramp/hooks/useBuy'
-import { useIsAvaxCSupported } from 'features/buyOnramp/hooks/useIsAvaxCSupported'
+import { useBuy } from 'features/meldOnramp/hooks/useBuy'
+import { useIsAvaxCSupported } from 'features/meldOnramp/hooks/useIsAvaxCSupported'
 import { useNavigateToSwap } from 'features/swap/hooks/useNavigateToSwap'
 import { useHasEnoughAvaxToStake } from 'hooks/earn/useHasEnoughAvaxToStake'
 import useStakingParams from 'hooks/earn/useStakingParams'
