--- packages/core-mobile/app/new/routes/(signedIn)/(modals)/buyOnramp/index.tsx
@@ -1,1 +1,0 @@
-export { BuyTokenScreen as default } from 'features/buyOnramp/screens/BuyTokenScreen'
