--- packages/core-mobile/app/new/routes/(signedIn)/(modals)/buyOnramp/selectBuyAmount.tsx
@@ -1,1 +1,0 @@
-export { SelectBuyAmountScreen as default } from 'features/buyOnramp/screens/SelectBuyAmountScreen'
