+++ packages/core-mobile/app/new/routes/(signedIn)/(modals)/meldOnramp/_layout.tsx
@@ -2,7 +2,7 @@ import { Stack } from 'common/components/Stack'
 import { useModalScreenOptions } from 'common/hooks/useModalScreenOptions'
 import React from 'react'
 
-export default function BuyOnrampLayout(): JSX.Element {
+export default function MeldOnrampLayout(): JSX.Element {
   const { modalStackNavigatorScreenOptions, modalFirstScreenOptions } =
     useModalScreenOptions()
   return (
