+++ packages/core-mobile/app/new/routes/(signedIn)/(modals)/meldOnramp/index.tsx
@@ -1,0 +1,1 @@
+export { BuyTokenScreen as default } from 'features/meldOnramp/screens/BuyTokenScreen'
