+++ packages/core-mobile/app/new/routes/(signedIn)/(modals)/meldOnramp/selectBuyAmount.tsx
@@ -1,0 +1,1 @@
+export { SelectBuyAmountScreen as default } from 'features/meldOnramp/screens/SelectBuyAmountScreen'
