+++ packages/core-mobile/app/new/routes/(signedIn)/(modals)/selectBuyToken/index.tsx
@@ -1,1 +1,1 @@
-export { SelectBuyTokenScreen as default } from 'features/buyOnramp/screens/SelectBuyTokenScreen'
+export { SelectBuyTokenScreen as default } from 'features/meldOnramp/screens/SelectBuyTokenScreen'
