+++ packages/core-mobile/app/new/routes/(signedIn)/(modals)/selectPaymentMethod/selectServiceProvider.tsx
@@ -1,1 +1,1 @@
-export { SelectServiceProviderScreen as default } from 'features/buyOnramp/screens/SelectServiceProviderScreen'
+export { SelectServiceProviderScreen as default } from 'features/meldOnramp/screens/SelectServiceProviderScreen'
