+++ packages/core-mobile/app/new/routes/(signedIn)/_layout.tsx
@@ -149,7 +149,7 @@ export default function WalletLayout(): JSX.Element {
             options={formSheetScreensOptions}
           />
           <Stack.Screen
-            name="(modals)/buyOnramp"
+            name="(modals)/meldOnramp"
             options={modalScreensOptions}
           />
           <Stack.Screen
